requirements.txt
joblib==1.0.1
nltk==3.6.3
pandas==1.3.3
streamlit==0.89.0
numpy==1.21.2
Pillow==8.3.2
scikit_learn==0.24.2
